Move Parameterized tests to selectors package
