+++ src/test/java/org/jenkinsci/plugins/runselector/selectors/ParameterizedRunSelectorTest.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package org.jenkinsci.plugins.runselector;
+package org.jenkinsci.plugins.runselector.selectors;
 
 import hudson.model.Cause;
 import hudson.model.FreeStyleProject;
@@ -33,8 +33,8 @@ import hudson.model.Run;
 import hudson.model.StringParameterDefinition;
 import hudson.model.StringParameterValue;
 import hudson.model.TaskListener;
+import org.jenkinsci.plugins.runselector.RunSelector;
 import org.jenkinsci.plugins.runselector.context.RunSelectorContext;
-import org.jenkinsci.plugins.runselector.selectors.ParameterizedRunSelector;
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.junit.ClassRule;
