+++ src/test/java/org/jenkinsci/plugins/runselector/selectors/RunSelectorParameterTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.jenkinsci.plugins.runselector;
+package org.jenkinsci.plugins.runselector.selectors;
 
 import com.gargoylesoftware.htmlunit.HttpMethod;
 import com.gargoylesoftware.htmlunit.WebClientOptions;
@@ -32,8 +32,6 @@ import hudson.cli.CLI;
 import hudson.model.FreeStyleProject;
 import hudson.model.ParametersDefinitionProperty;
 import hudson.model.Queue;
-import org.jenkinsci.plugins.runselector.selectors.RunSelectorParameter;
-import org.jenkinsci.plugins.runselector.selectors.StatusRunSelector;
 import org.junit.Rule;
 import org.junit.Test;
 import org.jvnet.hudson.test.CaptureEnvironmentBuilder;
